// src/app/pages/shared/interface.ts
export interface IEmployee {
    id: number;
    email: string;
    imageUrl: string;
    userName: string;
    repeatPassword: string;
    password: string;
    role: string;
    color: string[];
    status: number;
    country: string;
}
export interface ISelect {
    id?: string;
    name?: string;
}